Move everything actorlist related to its own file
