+++ scripts/actorStorage.js
@@ -1,3 +1,7 @@
+export const actorStorage = {
+    current: null,
+} // The currently selected actor
+
 export function saveLastActorId(actorId) {
     game.settings.set("sheet-only", "lastActorId", actorId);
 }
